refactor(policies): disambiguate lingbot_vla_v2 vendored base module names
Rename the two vendored base modules whose upstream names (without the _v2 suffix)
read as if they were the separate v1 policy:
  configuration_lingbot_vla.py -> configuration_lingbot_vla_v2_internal.py
  modeling_lingbot_vla.py      -> modeling_lingbot_vla_v2_base.py

These are internal building blocks of LingBot-VLA 2.0 (the HF config
LingbotVLAConfig/LingbotVLAV2Config and the shared AdaRMSNorm / QwenvlWithExpert /
FlowMatching(V1) base that FlowMatchingV2 is built on), not the v1 lingbot_va /
lingbot_vla policies (which live in their own directories). Update the relative
imports and the scoped ruff per-file-ignore glob accordingly. No behavioral change.
